# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NPR Seminar NestJS API
  version: 1.0.0
  description: |
    Authoritative contract for the single NestJS backend used by the NPR Seminar
    web and scanner clients.

    ADMIN and SCANNER actors use a Redis-backed project session cookie. Every
    state-changing session request requires X-CSRF-Token, an allowed same-origin
    Origin, and application/json. Public family access uses a separate,
    contact-bound X-Booking-Proof credential. QR tokens, booking proofs, OTPs,
    cookies, passwords, pairing codes, and full parent contact values are secrets and
    MUST be redacted from application, proxy, tracing, analytics, and audit logs.

    Durable resources are never hard-deleted. Booking, check-in, student, sync,
    circuit, and scanner-device histories are append-only and returned in stable
    chronological order.
  x-contract-authority: packages/contracts/openapi.yaml
  x-backend-runtime: NestJS modular monolith
  x-sensitive-log-redaction:
    - Cookie
    - Set-Cookie
    - X-CSRF-Token
    - X-Booking-Proof
    - bookingProof
    - X-QR-Token
    - qrToken
    - pairingCode
    - oneTimeCode
    - contact
    - motherPhone
    - fatherPhone
    - password
  x-canonical-student-snapshot:
    snapshotDate: 2026-07-17
    classNameAdmissionPolicy:
      normalizeForDecision: NFKC_TRIM
      preserveAcceptedSourceValue: true
      allowBracketless: true
      allowOneTerminalScheduleSuffix: true
...
      rejectAsterisk: true
...
    scope: ALL
...
servers:
...
paths:
...
  /api/v1/admin/sms/gateway-readiness:
    get:
      operationId: getSmsGatewayReadiness
      tags: [Admin SMS]
      summary: Inspect SMS gateway configuration and safety switches
      description: Returns only non-secret readiness flags; provider credentials and sender values are never exposed.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: SMS gateway readiness flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsGatewayReadiness'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/sms/templates:
    get:
      operationId: listSmsTemplates
      tags: [Admin SMS]
      summary: List SMS and LMS templates
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Templates ordered by active state and key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createSmsTemplate
      tags: [Admin SMS]
      summary: Create an SMS or LMS template
      description: Message type and byte counts are derived after NFC normalization and strict EUC-KR validation.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsTemplateCreateRequest'
      responses:
        '201':
          description: Template created with server-derived payload classification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTemplateCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/sms/templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/SmsTemplateId'
    get:
      operationId: getSmsTemplate
      tags: [Admin SMS]
      summary: Get one SMS or LMS template
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateSmsTemplate
      tags: [Admin SMS]
      summary: Update a template with optimistic version control
      description: The body version must equal the current decimal version string.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsTemplateUpdateRequest'
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: archiveSmsTemplate
      tags: [Admin SMS]
      summary: Archive a template by transitioning it to inactive
      description: No row is deleted; If-Match carries the current decimal version string.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      x-hard-delete: forbidden
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '200':
          description: Inactive template with incremented version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/sms/targets/preview:
    post:
      operationId: previewSmsTargets
      tags: [Admin SMS]
      summary: Preview a target snapshot and classified payload
      description: |
        This ephemeral calculation persists no message or recipient mutation.
        Exactly one of templateId or message is accepted. The seven supported
        variables are rendered independently for each recipient. At most ten
        masked, fully rendered samples are returned; full contact values never
        leave the server.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: ephemeral
      x-idempotency: exempt
      x-idempotency-exemption-reason: read-only-target-snapshot-calculation
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsTargetRequest'
      responses:
        '200':
          description: Current target snapshot and payload classification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsTargetPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/sms/sends:
    post:
      operationId: enqueueSmsSend
      tags: [Admin SMS]
      summary: Enqueue an administrator group SMS or LMS send
      description: The preview token must still match the exact recipients, session, template version, and rendered payloads or the request fails with 409.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsEnqueueRequest'
      responses:
        '202':
          description: Recipient messages durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsEnqueueAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/sms/survey-sends:
    post:
      operationId: enqueueSurveySmsSend
      tags: [Admin SMS]
      summary: Enqueue a survey SMS or LMS send
      description: Uses the same exact preview-token check as administrator group sends and records source SURVEY.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsEnqueueRequest'
      responses:
        '202':
          description: Survey recipient messages durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsEnqueueAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/sms/messages:
    get:
      operationId: listSmsMessages
      tags: [Admin SMS]
      summary: List recent SMS outbox messages newest first
      description: Returns masked recipients and safe provider metadata only; message bodies and full contacts are excluded.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SmsDeliveryStatus'
        - name: source
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SmsPurpose'
        - $ref: '#/components/parameters/BranchQuery'
        - name: seminarSessionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: batchId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Recent safe SMS history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsMessageList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/sms/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/SmsMessageId'
    get:
      operationId: getSmsMessage
      tags: [Admin SMS]
      summary: Get one SMS outbox message and append-only attempts
      description: Attempt events are immutable and never expose message bodies, credentials, or full recipients.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Safe message detail and chronological attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsMessageDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: npr_seminar_session
      description: |
        Opaque project-specific Redis session cookie for ADMIN and SCANNER.
        Production attributes are HttpOnly, Secure, SameSite=Lax, Path=/, with
        no Domain where possible. Cookie values are never logged.
    bookingProof:
      type: apiKey
      in: header
      name: X-Booking-Proof
      description: |
        Short-lived 256-bit base64url bearer credential bound to a normalized
        family contact digest, purpose-derived scopes, and expiry. PostgreSQL
        stores SHA-256(proof) only. The raw value is returned once, is independent
        from the project session cookie, and must be redacted from every log.
    qrCredential:
      type: apiKey
      in: header
      name: X-QR-Token
      description: |
        Opaque family QR credential used only for the minimal public pass.
        The raw value is secret, never persisted, and redacted from all logs.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated UUID scoped to actor and operation. Same key and same
        canonical request replays the stored secret-free result; same key with a
        different request returns 409. Raw one-time QR or pairing secrets are
        never present in replay snapshots.
      schema:
        type: string
        format: uuid
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Session-bound CSRF token. The value is secret and MUST NOT be logged.
      schema:
        type: string
        minLength: 32
        maxLength: 512
    Origin:
      name: Origin
      in: header
      required: true
      description: Exact configured browser origin; server also validates Host and trusted proxy metadata.
      schema:
        type: string
        format: uri
        maxLength: 300
...
    SmsTemplateId:
      name: templateId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SmsMessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: Current optimistic-lock version encoded as an unsigned decimal string.
      schema:
        type: string
        pattern: '^[0-9]+$'
        minLength: 1
        maxLength: 20
    BranchQuery:
      name: branch
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Branch'
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Malformed input, invalid header, or invalid confirmation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, invalid, or revoked session/proof/credential
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Role, CSRF, same-origin, contact ownership, or scope check failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found or intentionally concealed from this actor
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State transition, capacity, duplicate, lock, version, or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: Required PostgreSQL, Redis, source, or cryptographic configuration is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    SmsPurpose:
      type: string
      enum: [OTP, BOOKING_CONFIRMED, BOOKING_CANCELLED, FIRST_CHECK_IN, ADMIN_GROUP, SURVEY]
...
    SmsDeliveryStatus:
      type: string
      enum: [PENDING, CLAIMED, SENDING, SENT, BLOCKED_DISABLED, BLOCKED_ALLOWLIST, FAILED_PERMANENT, DELIVERY_UNKNOWN, DEAD, CANCELLED]
...
    SmsGatewayReadiness:
      type: object
      required: [enabled, configured, allowlistEnabled, testMode, adapterAvailable, workerOnly]
      properties:
        enabled:
          type: boolean
        configured:
          type: boolean
        allowlistEnabled:
          type: boolean
        testMode:
          type: boolean
        adapterAvailable:
          type: boolean
          description: Always false in the HTTP API process because the provider adapter is worker-isolated.
        workerOnly:
          type: boolean
          const: true
      additionalProperties: false
    SmsPayloadClassification:
      type: object
      required: [messageType, messageBytes, titleBytes]
      properties:
        messageType:
          $ref: '#/components/schemas/SmsMessageType'
        messageBytes:
          type: integer
          minimum: 1
          maximum: 2000
          description: NFC-normalized EUC-KR byte count.
        titleBytes:
          type: [integer, 'null']
          minimum: 1
          maximum: 44
          description: NFC-normalized EUC-KR byte count; null when no LMS title exists.
      additionalProperties: false
    SmsTemplate:
      type: object
      required: [templateId, key, name, purpose, title, body, active, version, createdAt, updatedAt]
      properties:
        templateId:
          type: string
          format: uuid
        key:
          type: string
          pattern: '^[A-Z0-9_]{3,80}$'
        name:
          type: string
          minLength: 1
          maxLength: 160
        purpose:
          $ref: '#/components/schemas/SmsPurpose'
        title:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: Server additionally enforces at most 44 EUC-KR bytes and LMS-only use.
        body:
          type: string
          minLength: 1
          maxLength: 2000
          description: Server additionally enforces 1..2000 EUC-KR bytes.
        active:
          type: boolean
        version:
          type: string
          pattern: '^[0-9]+$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    SmsTemplateList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SmsTemplate'
      additionalProperties: false
    SmsTemplateCreateRequest:
      type: object
      required: [key, name, purpose, body]
      properties:
        key:
          type: string
          pattern: '^[A-Z0-9_]{3,80}$'
        name:
          type: string
          minLength: 1
          maxLength: 160
        purpose:
          $ref: '#/components/schemas/SmsPurpose'
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Accepted only for a payload classified as LMS and limited to 44 EUC-KR bytes.
        body:
          type: string
          minLength: 1
          maxLength: 2000
          description: Limited to 2000 EUC-KR bytes after NFC normalization.
      additionalProperties: false
    SmsTemplateUpdateRequest:
      type: object
      required: [version]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        purpose:
          $ref: '#/components/schemas/SmsPurpose'
        title:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: Null removes the title; non-null titles require LMS and at most 44 EUC-KR bytes.
        body:
          type: string
          minLength: 1
          maxLength: 2000
        active:
          type: boolean
        version:
          type: string
          pattern: '^[0-9]+$'
      additionalProperties: false
    SmsTemplateCreated:
      type: object
      required: [templateId, key, name, purpose, title, body, active, version, createdAt, updatedAt, classification]
      properties:
        templateId:
          type: string
          format: uuid
        key:
          type: string
          pattern: '^[A-Z0-9_]{3,80}$'
        name:
          type: string
          minLength: 1
          maxLength: 160
        purpose:
          $ref: '#/components/schemas/SmsPurpose'
        title:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 2000
        active:
          type: boolean
        version:
          type: string
          pattern: '^[0-9]+$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        classification:
          $ref: '#/components/schemas/SmsPayloadClassification'
      additionalProperties: false
    SmsTargetRequest:
      type: object
      required: [branch, seminarSessionId, audience]
      properties:
        branch:
          $ref: '#/components/schemas/Branch'
        seminarSessionId:
          type: string
          format: uuid
        audience:
          $ref: '#/components/schemas/SmsAudience'
        templateId:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: Template text supporting {학생명}, {설명회명}, {일시}, {장소}, {QR링크}, {문의전화}, and {설문링크}.
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Optional title template; each rendered recipient payload must classify as LMS and remain within 44 EUC-KR bytes.
      oneOf:
        - required: [templateId]
          not:
            anyOf:
              - required: [message]
              - required: [title]
        - required: [message]
          not:
            required: [templateId]
      additionalProperties: false
    SmsEnqueueRequest:
      type: object
      required: [branch, seminarSessionId, audience, previewToken]
      properties:
        branch:
          $ref: '#/components/schemas/Branch'
        seminarSessionId:
          type: string
          format: uuid
        audience:
          $ref: '#/components/schemas/SmsAudience'
        templateId:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          maxLength: 2000
        title:
          type: string
          minLength: 1
          maxLength: 200
        previewToken:
          type: string
          minLength: 20
          maxLength: 100
      oneOf:
        - required: [templateId]
          not:
            anyOf:
              - required: [message]
              - required: [title]
        - required: [message]
          not:
            required: [templateId]
      additionalProperties: false
...
    SmsTargetPreview:
      type: object
      required: [branch, seminarSessionId, audience, recipientCount, previewToken, maskedRecipients, templateId, templateName, messageTemplate, titleTemplate, samples, maximumMessageType, maximumMessageBytes, maximumTitleBytes]
      properties:
        branch:
          $ref: '#/components/schemas/Branch'
        seminarSessionId:
          type: string
          format: uuid
        audience:
          $ref: '#/components/schemas/SmsAudience'
        recipientCount:
          type: integer
          minimum: 0
        previewToken:
          type: string
          minLength: 20
          maxLength: 100
        maskedRecipients:
          type: array
          maxItems: 10
          items:
            type: string
            pattern: '^\*\*\*-\*\*\*\*-[0-9]{4}$'
        templateId:
          type: [string, 'null']
          format: uuid
        templateName:
          type: string
          minLength: 1
          maxLength: 160
        messageTemplate:
          type: string
          minLength: 1
          maxLength: 2000
        titleTemplate:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
        samples:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/SmsRenderedSample'
        maximumMessageType:
          type: [string, 'null']
          enum: [SMS, LMS, null]
        maximumMessageBytes:
          type: [integer, 'null']
          minimum: 1
          maximum: 2000
        maximumTitleBytes:
          type: [integer, 'null']
          minimum: 1
          maximum: 44
      additionalProperties: false
    SmsEnqueueAccepted:
      type: object
      required: [batchId, queuedCount, previewToken, status, source, templateId, templateName]
      properties:
        batchId:
          type: string
          format: uuid
        queuedCount:
          type: integer
          minimum: 0
        previewToken:
          type: string
          minLength: 20
          maxLength: 100
        status:
          type: string
          const: QUEUED
        source:
          type: string
          enum: [ADMIN_GROUP, SURVEY]
        templateId:
          type: [string, 'null']
          format: uuid
        templateName:
          type: string
          minLength: 1
          maxLength: 160
      additionalProperties: false
...
    SmsMessageList:
      type: object
      required: [batches, items]
      properties:
        batches:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/SmsBatchSummary'
        items:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/SmsMessageSummary'
      additionalProperties: false
...
    SmsMessageDetail:
      type: object
      required: [messageId, batchId, source, branch, seminarSessionId, familyBookingId, templateId, templateName, audience, maskedRecipient, messageType, messageBytes, status, attemptCount, providerMessageId, providerResultCode, lastErrorCode, actorSubject, createdAt, updatedAt, attempts]
      properties:
        messageId:
          type: string
          format: uuid
        batchId:
          type: [string, 'null']
          format: uuid
        source:
          $ref: '#/components/schemas/SmsPurpose'
        branch:
          $ref: '#/components/schemas/Branch'
        seminarSessionId:
          type: [string, 'null']
          format: uuid
        familyBookingId:
          type: [string, 'null']
          format: uuid
        templateId:
          type: [string, 'null']
          format: uuid
        templateName:
          type: [string, 'null']
          maxLength: 160
        audience:
          type: [string, 'null']
          enum: [BOOKED_FAMILIES, RESERVED_FAMILIES, CHECKED_IN_FAMILIES, CANCELLED_FAMILIES, null]
        maskedRecipient:
          type: string
          pattern: '^\*\*\*-\*\*\*\*-[0-9]{4}$'
        messageType:
          $ref: '#/components/schemas/SmsMessageType'
        messageBytes:
          type: integer
          minimum: 1
          maximum: 2000
        status:
          $ref: '#/components/schemas/SmsDeliveryStatus'
        attemptCount:
          type: integer
          minimum: 0
          maximum: 12
        providerMessageId:
          type: [string, 'null']
          maxLength: 80
        providerResultCode:
          type: [integer, 'null']
        lastErrorCode:
          type: [string, 'null']
          maxLength: 80
        actorSubject:
          type: [string, 'null']
          maxLength: 160
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/SmsAttempt'
      additionalProperties: false
...